Added Error 404 exceptions where needed

+++ src/main/java/com/svalero/euroapi/controller/ArtistController.java
@@ -1,16 +1,29 @@
 package com.svalero.euroapi.controller;
 import com.svalero.euroapi.domain.Artist;
+import com.svalero.euroapi.domain.ErrorResponse;
+import com.svalero.euroapi.exception.ArtistNotFoundException;
+import com.svalero.euroapi.exception.SongNotFoundException;
 import com.svalero.euroapi.service.ArtistService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+import java.util.Optional;
 
 @RestController
 public class ArtistController {
 
     @Autowired
     private ArtistService artistService;
+
+    @GetMapping("/artist/{artistId}")
+    public Artist getArtist(@PathVariable long artistId) throws ArtistNotFoundException {
+        Optional<Artist> optionalArtist = artistService.getArtistById(artistId);
+        return optionalArtist.orElseThrow(() -> new ArtistNotFoundException(artistId));
+    }
+
 
     @GetMapping("/artists")
     public List<Artist> getAll(@RequestParam(defaultValue = "") String name,
@@ -44,11 +57,27 @@ public class ArtistController {
         artistService.saveArtist(artist);
     }
     @DeleteMapping("/artist/{artistId}")
-    public void removeArtist(@PathVariable long artistId) {
-        artistService.removeArtist(artistId);
+    public void removeArtist(@PathVariable long artistId) throws ArtistNotFoundException {
+        Optional<Artist> optionalArtist = artistService.getArtistById(artistId);
+            if(optionalArtist.isPresent()) {
+                artistService.removeArtist(artistId);
+            } else {
+                throw new ArtistNotFoundException(artistId);
+            }
     }
     @PutMapping("/artist/{artistId}")
-    public void modifyArtist(@RequestBody Artist artist, @PathVariable long artistId) {
-        artistService.modifyArtist(artist, artistId);
+    public void modifyArtist(@RequestBody Artist artist, @PathVariable long artistId) throws ArtistNotFoundException {
+        Optional<Artist> optionalArtist = artistService.getArtistById(artistId);
+        if (optionalArtist.isPresent()) {
+            artistService.modifyArtist(artist, artistId);
+        } else {
+            throw new ArtistNotFoundException(artistId);
+        }
     }
+    @ExceptionHandler(ArtistNotFoundException.class)
+    public ResponseEntity<ErrorResponse> artistNotFoundException(ArtistNotFoundException anfe) {
+        ErrorResponse errorResponse = new ErrorResponse(404, anfe.getMessage());
+        return new ResponseEntity<>(errorResponse, HttpStatus.NOT_FOUND);
+    }
+
 }
