+++ src/main/java/com/svalero/euroapi/controller/CountryController.java
@@ -1,16 +1,25 @@
 package com.svalero.euroapi.controller;
 import com.svalero.euroapi.domain.Country;
-import com.svalero.euroapi.domain.Song;
+import com.svalero.euroapi.domain.ErrorResponse;
+import com.svalero.euroapi.exception.CountryNotFoundException;
 import com.svalero.euroapi.service.CountryService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import java.util.List;
+import java.util.Optional;
 
 @RestController
 public class CountryController {
     @Autowired
     private CountryService countryService;
 
+    @GetMapping("/country/{countryId}")
+    public Country getCountry(@PathVariable long countryId) throws CountryNotFoundException {
+        Optional<Country> optionalCountry = countryService.getCountryById(countryId);
+        return optionalCountry.orElseThrow(() -> new CountryNotFoundException(countryId));
+    }
     @GetMapping("/countries")
     public List<Country> getAll(@RequestParam(defaultValue = "") String countryName,
                                 @RequestParam(defaultValue = "") String bigFive,
@@ -44,12 +53,28 @@ public class CountryController {
     }
 
     @DeleteMapping("/country/{countryId}")
-    public void removeCountry(@PathVariable long countryId) {
-        countryService.removeCountry(countryId);
+    public void removeCountry(@PathVariable long countryId) throws CountryNotFoundException {
+        Optional<Country> optionalCountry = countryService.getCountryById(countryId);
+            if(optionalCountry.isPresent()) {
+                countryService.removeCountry(countryId);
+            } else {
+                throw new CountryNotFoundException(countryId);
+            }
     }
 
     @PutMapping("/country/{countryId}")
-    public void modifyCountry(@RequestBody Country country, @PathVariable long countryId) {
-        countryService.modifyCountry(country, countryId);
+    public void modifyCountry(@RequestBody Country country, @PathVariable long countryId) throws CountryNotFoundException {
+        Optional<Country> optionalCountry = countryService.getCountryById(countryId);
+        if (optionalCountry.isPresent()) {
+            countryService.modifyCountry(country, countryId);
+        } else {
+            throw new CountryNotFoundException(countryId);
+        }
+    }
+
+    @ExceptionHandler(CountryNotFoundException.class)
+    public ResponseEntity<ErrorResponse> countryNotFoundException(CountryNotFoundException cnfe) {
+        ErrorResponse errorResponse = new ErrorResponse(404, cnfe.getMessage());
+        return new ResponseEntity<>(errorResponse, HttpStatus.NOT_FOUND);
     }
 }
